acomodando valores para cable, valores flotantes

+++ app/Models/VentaProducto.php
@@ -27,7 +27,7 @@ class VentaProducto extends Model
         return $this->belongsTo(Venta::class);
     }
 
-    public static function validateVentaProducto(array $data, bool $sumarCantidad = true): false|int
+    public static function validateVentaProducto(array $data, bool $sumarCantidad = true): false|float
     {
         $productoActual = Producto::find($data['producto_id']);
         $cantidadRequerida = $data['cantidad'];
@@ -64,11 +64,13 @@ class VentaProducto extends Model
             'precio' => $data['precio'],
         ]);
 
+
         $ventaTotal = self::where('venta_id', $data['venta_id'])
             ->get()
-            ->sum(fn ($item) => $item->cantidad * $item->precio);
+            ->sum(fn($item) => $item->cantidad * $item->precio);
 
-        Venta::where('id', $data['venta_id'])->update(['venta_total' => $ventaTotal]);
+        Venta::where('id', $data['venta_id'])
+            ->update(['venta_total' => $ventaTotal]);
 
         return $ventaProducto;
     }
@@ -143,11 +145,11 @@ class VentaProducto extends Model
                     ->orderBy('created_at', $order_date);
             })
             ->get()
-            ->groupBy(fn ($item) => $item->producto->categoria->nombre)
+            ->groupBy(fn($item) => $item->producto->categoria->nombre)
             ->map(function ($item, $categoria) {
                 return [
                     'categoria' => $categoria,
-                    'total' => $item->sum(fn ($i) => $i->cantidad * $i->precio),
+                    'total' => $item->sum(fn($i) => $i->cantidad * $i->precio),
                 ];
             })->values();
     }
